refactor(ui): move every dialoguer menu into ui::menus
Connected apps, the MCP/app catalog, webhook channels, skills, personas, and the
Telegram/Discord integrations — 1,514 lines of arrow-key menus sitting in main.rs
next to the REPL loop they suspend.

They belong together for a reason that is not filing: every one of them drives
the terminal directly, which is exactly the property the retained frame keys off
(`stdin: Stdin::Always` on their row in the slash table). Having them in one
module makes that shared constraint visible instead of implied.

Cut cost: 8 helpers widened to `pub(crate)`, 4 exported back, 3 re-imported by
the slash handlers. Two clippy suggestions (`split_once`, `sort_by_key`) travel
with the code unchanged — they predate this and fixing them here would hide a
behaviour change inside a move.

main.rs: 9,062 -> 7,546
